Debt Payoff with Monthly Budget

+++ pkg/model/tradeline.go
@@ -4,8 +4,8 @@ package model
 type Tradeline struct {
 	ID                      string  `json:"id"`
 	AccountName             string  `json:"account_name"`
-	Balance                 float64 `json:"balance"`
-	MinimumPayment          float64 `json:"minimum_payment"`
+	Balance                 int     `json:"balance"`
+	MinimumPayment          int     `json:"minimum_payment"`
 	InterestRate            float64 `json:"interest_rate"`
 	RevolvingOrInstallement string  `json:"revolving_or_installement"`
 }
